Move the audit API to its own header files.
